requirements.txt
asgiref==3.2.3
attrs==19.3.0
certifi==2019.11.28
chardet==3.0.4
Click==7.0
Django==3.0.3
django-pytest==0.2.0
djangorestframework==3.10.3
entrypoints==0.3
flake8==3.7.9
Flask==1.1.1
idna==2.8
importlib-metadata==1.5.0
itsdangerous==1.1.0
Jinja2==2.11.1
MarkupSafe==1.1.1
mccabe==0.6.1
more-itertools==8.2.0
packaging==20.1
pluggy==0.13.1
psycopg2-binary==2.8.4
py==1.8.1
pycodestyle==2.5.0
pyflakes==2.1.1
pyparsing==2.4.6
pytest==5.3.5
pytz==2019.3
requests==2.22.0
six==1.14.0
sqlparse==0.3.0
urllib3==1.25.8
wcwidth==0.1.8
Werkzeug==1.0.0
zipp==2.2.0

